deploy/requirements.txt
yfinance==0.2.43
google-cloud-storage==2.18.2
scikit-learn==1.3.0
joblib==1.3.0
numpy==1.24.3